backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
openai>=1.12.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.7.0
python-multipart>=0.0.9
weasyprint>=62.0
jinja2>=3.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.27.0
sqlalchemy>=2.0.0
